Added api call to let a user leave a chat
